requirements.txt
# Streamlit and core dependencies
streamlit>=1.28.0

# Backend dependencies
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
python-multipart>=0.0.6
pypdf2>=3.0.1
requests>=2.31.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pillow>=10.0.0
python-docx>=1.2.0
lxml>=6.0.0
reportlab>=4.0.0
